feat: ローディングの状態管理のコンテキストを追加しGlobalHandlerから該当処理を削除
